Move real data test and benchmark to one file
